+++ packages/isar_test/lib/common.dart
@@ -71,7 +71,8 @@ void registerBinaries() {
       Isar.initializeLibraries(
         libraries: {
           'windows': path.join(dartToolDir, 'libisar_windows_x64.dll'),
-          'macos': path.join(dartToolDir, 'libisar_macos_x64.dylib'),
+          'macos': path.join(dartToolDir,
+              '/Users/simon/Documents/GitHub/isar-core/dart-ffi/target/x86_64-apple-darwin/debug/libisar.dylib'),
           'linux': path.join(dartToolDir, 'libisar_linux_x64.so'),
         },
       );