+++ packages/isar_test/test_driver/isar.dart
@@ -30,6 +30,6 @@ void executeTests(Completer<bool> completer) {
       completer.complete(testCount != 0 && allTestsSuccessful);
     });
 
-    tests.run();
+    tests.main();
   });
 }